+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css'
 import {useEffect, useState} from "react";
 import {BrowserRouter as Router, Route, Routes} from "react-router-dom";
-import {getMenu} from "./puzzleApi.ts";
+import {getMenu} from "./PuzzleApi.ts";
 import {Category, FixedEncodingData, Level, Menu, Puzzle, VariableEncodingData} from "./Menu.ts";
 import MenuDisplay from "./components/MenuDisplay.tsx";
 import LevelMenu from "./components/LevelMenu.tsx";
